Set authenticated user on global context provider render
